advances on tp4

+++ backend/main/resources/poem.py
@@ -4,19 +4,26 @@ from datetime import datetime
 from sqlalchemy import func
 from .. import db
 from main.models import PoemModel, UserModel, ScoreModel
+from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt
 
 
 class Poem(Resource):
 
+    @jwt_required()
     def get(self, id):
         poem = db.session.query(PoemModel).get_or_404(id)
         return poem.to_json()
 
+    @jwt_required()
     def delete(self, id):
+        claims = get_jwt()
+        user_ident = get_jwt_identity()
         poem = db.session.query(PoemModel).get_or_404(id)
-        db.session.delete(poem)
-        db.session.commit()
-        return '', 204
+        if 'role' in claims:
+            if claims['role'] == 'admin' or int(poem.user_id) == user_ident:
+                db.session.delete(poem)
+                db.session.commit()
+                return '', 204
 
 
 class Poems(Resource):
@@ -26,48 +33,75 @@ class Poems(Resource):
         page = 1
         per_page = 10
         poems = db.session.query(PoemModel)
-        if request.get_json():
-            filters = request.get_json().items()
-            for key, value in filters:
-                if key == 'page':
-                    page = int(value)
-                if key == 'per_page':
-                    per_page = int(value)
-                if key == 'title':
-                    poems = poems.filter(PoemModel.title.like('%' + value + '%'))
-                if key == 'userId':
-                    poems = poems.filter(PoemModel.userId == value)
-                if key == 'date[gte]':
-                    poems = poems.filter(PoemModel.date >= datetime.strptime(value, '%d-%m-%Y'))
-                if key == 'date[lte]':
-                    poems = poems.filter(PoemModel.date <= datetime.strptime(value, '%d-%m-%Y'))
-                if key == 'author':
-                    poems = poems.filter(PoemModel.user.has(UserModel.name.like('%' + value + '%')))
+        claims = get_jwt()
+        user_ident = get_jwt_identity()
+        if user_ident:
+            if request.get_json():
+                filters = request.get_json().items()
+                for key,value in filters:
+                    if key == 'page':
+                        page = int(value)
+                    if key == 'per_page':
+                        per_page = int(value)
 
-                if key == 'sort_by':
-                    if value == 'date':
-                        poems = poems.order_by(PoemModel.date)
-                    if value == 'date[desc]':
-                        poems = poems.order_by(PoemModel.date.desc())
-                    if value == 'score':
-                        poems = poems.outerjoin(PoemModel.scores).group_by(PoemModel.id).order_by(func.avg(ScoreModel.score))
-                    if value == 'score[desc]':
-                        poems = poems.outerjoin(PoemModel.scores).group_by(PoemModel.id).order_by(func.avg(ScoreModel.score).desc())
-                    if value == 'author':
-                        poems = poems.outerjoin(PoemModel.user).group_by(PoemModel.id).order_by(UserModel.name)
-                    if value == 'author[desc]':
-                        poems = poems.outerjoin(PoemModel.user).group_by(PoemModel.id).order_by(UserModel.name.desc())
+            poems = db.session.query(PoemModel).filter(PoemModel.user_id != user_ident)
+            poems = poems.outerjoin(PoemModel.scores).group_by(PoemModel.id).order_by(func.count(PoemModel.scores))
+
+        else:
+            if request.get_json():
+                filters = request.get_json().items()
+                for key, value in filters:
+                    if key == 'page':
+                        page = int(value)
+                    if key == 'per_page':
+                        per_page = int(value)
+                    if key == 'title':
+                        poems = poems.filter(PoemModel.title.like('%' + value + '%'))
+                    if key == 'userId':
+                        poems = poems.filter(PoemModel.userId == value)
+                    if key == 'date[gte]':
+                        poems = poems.filter(PoemModel.date >= datetime.strptime(value, '%d-%m-%Y'))
+                    if key == 'date[lte]':
+                        poems = poems.filter(PoemModel.date <= datetime.strptime(value, '%d-%m-%Y'))
+                    if key == 'author':
+                        poems = poems.filter(PoemModel.user.has(UserModel.name.like('%' + value + '%')))
+
+                    if key == 'sort_by':
+                        if value == 'date':
+                            poems = poems.order_by(PoemModel.date)
+                        if value == 'date[desc]':
+                            poems = poems.order_by(PoemModel.date.desc())
+                        if value == 'score':
+                            poems = poems.outerjoin(PoemModel.scores).group_by(PoemModel.id).order_by(func.avg(ScoreModel.score))
+                        if value == 'score[desc]':
+                            poems = poems.outerjoin(PoemModel.scores).group_by(PoemModel.id).order_by(func.avg(ScoreModel.score).desc())
+                        if value == 'author':
+                            poems = poems.outerjoin(PoemModel.user).group_by(PoemModel.id).order_by(UserModel.name)
+                        if value == 'author[desc]':
+                            poems = poems.outerjoin(PoemModel.user).group_by(PoemModel.id).order_by(UserModel.name.desc())
 
         poems = poems.paginate(page, per_page, True, 10)
-        return jsonify({"poems": [poem.to_json_short() for poem in poems.items],
-                        "total": poems.total,
-                        "pages": poems.pages,
-                        "page": page})
+        if 'role' in claims:
+            if claims['role'] == 'admin':
+                return jsonify({"poems": [poem.to_json_short() for poem in poems.items],
+                                "total": poems.total,
+                                "pages": poems.pages,
+                                "page": page})
 
-    # @jwt_required
+    @jwt_required
     def post(self):
-        poems = PoemModel.from_json(request.get_json())
-        db.session.add(poems)
-        db.session.commit()
-        return poems.to_json(), 201
+        user_ident = get_jwt_identity()
+        poem = PoemModel.from_json(request.get_json())
+        user = db.session.query(UserModel).get_or_404(user_ident)
+        claims = get_jwt()
+        if 'role' in claims and claims['role'] == 'poet':
+            if len(user.poems) == 0 or len(user.scores) >= 2:
+                poem.user_id = user_ident
+                db.session.add(poem)
+                db.session.commit()
+                return poem.to_json(), 201
+            else:
+                return 'There are not enough reviews from this user', 400
+        else:
+            return 'You are not a poet, only poets can create poems', 401
